add upc search

+++ app/src/main/java/au/com/holberton/simplicity/mobile/productdetails/ListingDetailsRepository.kt
@@ -3,11 +3,11 @@ package au.com.holberton.simplicity.mobile.productlisting
 import au.com.holberton.simplicity.mobile.productdetails.ListingDetails
 
 object ListingDetailsRepository {
-    suspend fun getListingDetails(upc: Double): ListingDetails {
+    suspend fun getListingDetails(upc: Long): ListingDetails {
         return ListingApi.service.getListingDetails(upc)
     }
 
-    suspend fun updateItemQty(upc: Double, quantity: Int?): Unit {
+    suspend fun updateItemQty(upc: Long, quantity: Int?): Unit {
         ListingApi.service.updateQuantity(upc, quantity)
     }
 }
